App convertida en single activity. La LoginActivity era android:windowSoftInputMode="adjustPan", pero no se como aplicarlo al LoginFragment. Los tests de instrumentación deberian modificarse.
